fix BigQuery validation
