+++ scio-bigquery/src/test/scala/com/spotify/scio/bigquery/validation/BigQueryValidationTest.scala
@@ -26,12 +26,6 @@ import org.scalatest.{BeforeAndAfterAll, FlatSpec, Matchers}
 // properties on the individual field level processing data
 class BigQueryValidationTest extends FlatSpec with Matchers with BeforeAndAfterAll  {
 
-  override def beforeAll(): Unit = {
-    // We need this at runtime as well and tests are run in a fork
-    SampleOverrideTypeProvider.setSystemProperty()
-  }
-
-  @SampleOverrideTypeProvider.setProperty
   @BigQueryType.fromSchema(
     """
       |{
@@ -44,7 +38,6 @@ class BigQueryValidationTest extends FlatSpec with Matchers with BeforeAndAfterA
     """.stripMargin)
   class CountryInput
 
-  @SampleOverrideTypeProvider.setProperty
   @BigQueryType.toTable
   case class CountryOutput(@description("COUNTRY") country: Country,
                            countryString: String,
